Added all files from Campus GIt Server
Personally built files,

Backend.java - This is the main backend class that creates the graph and allows the frontend to access multiple features

BackendDeveloperTests.java - This tests the backend implementation in three ways.

BackendInterface.java - This is the interface for the backend file that describes the methods implemented by the backend.

CS400Graph.java - This graph implementation uses a modified Vertex class that includes one extra variable, that's needed in the backend implementation.

I was the backend Developer on the team.